Exception handling - Delete

+++ src/main/java/com/educandoweb/springAula/services/UserService.java
@@ -4,10 +4,13 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.educandoweb.springAula.entities.User;
 import com.educandoweb.springAula.repositories.UserRepository;
+import com.educandoweb.springAula.services.exceptions.DatabaseException;
 import com.educandoweb.springAula.services.exceptions.ResourceNotFoundException;
 
 // @Component // REGISTRA A CLASSE COMO COMPONENTE DO SPRING E PODERA SER INJETADO AUTOMATICAMENTE PELO @AUTOWIRED //
@@ -32,8 +35,15 @@ public class UserService {
 	}
 	
 	//DELEÇÃO DO USUARIO
-	public void delete(Long id) {
-		repository.deleteById(id);
+	public void delete(Long id){
+	    try {
+	        if(!repository.existsById(id)) throw new ResourceNotFoundException(id);
+	        repository.deleteById(id);
+	    }catch (EmptyResultDataAccessException e){
+	        throw new ResourceNotFoundException(id);
+	    }catch(DataIntegrityViolationException e) {
+	    	throw new DatabaseException(e.getMessage());
+	    }
 	}
 	
 	//ATUALIZAR O USUARIO
